feat(fast-element): add observation of context properties and fix lint
