+++ packages/web-components/fast-element/src/directives/behavior.ts
@@ -1,4 +1,4 @@
-import { ExecutionContext } from "src/interfaces";
+import { ExecutionContext } from "../observation/observable";
 
 export interface Behavior {
     bind(source: unknown, context: ExecutionContext): void;
