+++ packages/web-components/fast-element/src/view.ts
@@ -1,5 +1,5 @@
 import { Behavior } from "./directives/behavior";
-import { ExecutionContext } from "./interfaces";
+import { ExecutionContext } from "./observation/observable";
 
 /**
  * Represents a collection of DOM nodes which can be bound to a data source.
