+++ packages/web-components/fast-element/tsconfig.json
@@ -4,7 +4,8 @@
         "outDir": "dist",
         "baseUrl": ".",
         "strictNullChecks": true,
-        "target": "es2015"
+        "target": "es2015",
+        "experimentalDecorators": true
     },
     "include": [
         "src/**/*"
